refactor selector
