+++ src/meta/src/hummock/compaction/level_selector.rs
@@ -24,12 +24,11 @@ use risingwave_pb::hummock::hummock_version::Levels;
 use risingwave_pb::hummock::CompactionConfig;
 
 use crate::hummock::compaction::compaction_config::CompactionConfigBuilder;
-use crate::hummock::compaction::manual_compaction_picker::ManualCompactionPicker;
 use crate::hummock::compaction::min_overlap_compaction_picker::MinOverlappingPicker;
 use crate::hummock::compaction::overlap_strategy::OverlapStrategy;
 use crate::hummock::compaction::{
     create_overlap_strategy, CompactionInput, CompactionPicker, CompactionTask,
-    LevelCompactionPicker, ManualCompactionOption, TierCompactionPicker,
+    LevelCompactionPicker, TierCompactionPicker,
 };
 use crate::hummock::level_handler::LevelHandler;
 
@@ -45,33 +44,28 @@ pub trait LevelSelector: Sync + Send {
         level_handlers: &mut [LevelHandler],
     ) -> Option<CompactionTask>;
 
-    fn manual_pick_compaction(
-        &self,
-        task_id: HummockCompactionTaskId,
-        levels: &Levels,
-        level_handlers: &mut [LevelHandler],
-        option: ManualCompactionOption,
-    ) -> Option<CompactionTask>;
-
     fn name(&self) -> &'static str;
 }
 
 #[derive(Default)]
 pub struct SelectContext {
-    level_max_bytes: Vec<u64>,
+    pub level_max_bytes: Vec<u64>,
 
     // All data will be placed in the last level. When the cluster is empty, the files in L0 will
     // be compact to `max_level`, and the `max_level` would be `base_level`. When the total
     // size of the files in  `base_level` reaches its capacity, we will place data in a higher
     // level, which equals to `base_level -= 1;`.
-    base_level: usize,
-    score_levels: Vec<(u64, usize, usize)>,
+    pub base_level: usize,
+    pub score_levels: Vec<(u64, usize, usize)>,
 }
 
-// TODO: Set these configurations by meta rpc
-pub struct DynamicLevelSelector {
+pub struct LevelSelectorCore {
     config: Arc<CompactionConfig>,
     overlap_strategy: Arc<dyn OverlapStrategy>,
+}
+
+pub struct DynamicLevelSelector {
+    inner: LevelSelectorCore,
 }
 
 impl Default for DynamicLevelSelector {
@@ -84,10 +78,26 @@ impl Default for DynamicLevelSelector {
 
 impl DynamicLevelSelector {
     pub fn new(config: Arc<CompactionConfig>, overlap_strategy: Arc<dyn OverlapStrategy>) -> Self {
-        DynamicLevelSelector {
+        Self {
+            inner: LevelSelectorCore::new(config, overlap_strategy),
+        }
+    }
+}
+
+impl LevelSelectorCore {
+    pub fn new(config: Arc<CompactionConfig>, overlap_strategy: Arc<dyn OverlapStrategy>) -> Self {
+        Self {
             config,
             overlap_strategy,
         }
+    }
+
+    pub fn get_config(&self) -> &CompactionConfig {
+        self.config.as_ref()
+    }
+
+    pub fn get_overlap_strategy(&self) -> Arc<dyn OverlapStrategy> {
+        self.overlap_strategy.clone()
     }
 
     fn create_compaction_picker(
@@ -123,7 +133,7 @@ impl DynamicLevelSelector {
     /// `calculate_level_base_size` calculate base level and the base size of LSM tree build for
     /// current dataset. In other words,  `level_max_bytes` is our compaction goal which shall
     /// reach. This algorithm refers to the implementation in  `</>https://github.com/facebook/rocksdb/blob/v7.2.2/db/version_set.cc#L3706</>`
-    fn calculate_level_base_size(&self, levels: &Levels) -> SelectContext {
+    pub fn calculate_level_base_size(&self, levels: &Levels) -> SelectContext {
         let mut first_non_empty_level = 0;
         let mut max_level_size = 0;
         let mut ctx = SelectContext::default();
@@ -239,7 +249,11 @@ impl DynamicLevelSelector {
         ctx
     }
 
-    fn create_compaction_task(&self, input: CompactionInput, base_level: usize) -> CompactionTask {
+    pub fn create_compaction_task(
+        &self,
+        input: CompactionInput,
+        base_level: usize,
+    ) -> CompactionTask {
         let target_file_size = if input.target_level == 0 {
             self.config.target_file_size_base
         } else {
@@ -263,7 +277,7 @@ impl DynamicLevelSelector {
 
 impl LevelSelector for DynamicLevelSelector {
     fn need_compaction(&self, levels: &Levels, level_handlers: &[LevelHandler]) -> bool {
-        let ctx = self.get_priority_levels(levels, level_handlers);
+        let ctx = self.inner.get_priority_levels(levels, level_handlers);
         ctx.score_levels
             .first()
             .map(|(score, _, _)| *score > SCORE_BASE)
@@ -276,45 +290,20 @@ impl LevelSelector for DynamicLevelSelector {
         levels: &Levels,
         level_handlers: &mut [LevelHandler],
     ) -> Option<CompactionTask> {
-        let ctx = self.get_priority_levels(levels, level_handlers);
+        let ctx = self.inner.get_priority_levels(levels, level_handlers);
         for (score, select_level, target_level) in ctx.score_levels {
             if score <= SCORE_BASE {
                 return None;
             }
-            let picker = self.create_compaction_picker(select_level, target_level);
+            let picker = self
+                .inner
+                .create_compaction_picker(select_level, target_level);
             if let Some(ret) = picker.pick_compaction(levels, level_handlers) {
                 ret.add_pending_task(task_id, level_handlers);
-                return Some(self.create_compaction_task(ret, ctx.base_level));
+                return Some(self.inner.create_compaction_task(ret, ctx.base_level));
             }
         }
         None
-    }
-
-    fn manual_pick_compaction(
-        &self,
-        task_id: HummockCompactionTaskId,
-        levels: &Levels,
-        level_handlers: &mut [LevelHandler],
-        option: ManualCompactionOption,
-    ) -> Option<CompactionTask> {
-        let ctx = self.get_priority_levels(levels, level_handlers);
-        let target_level = if option.level == 0 {
-            ctx.base_level
-        } else if option.level == self.config.max_level as usize {
-            option.level
-        } else {
-            option.level + 1
-        };
-        if option.level > 0 && option.level < ctx.base_level {
-            return None;
-        }
-
-        let picker =
-            ManualCompactionPicker::new(self.overlap_strategy.clone(), option, target_level);
-
-        let ret = picker.pick_compaction(levels, level_handlers)?;
-        ret.add_pending_task(task_id, level_handlers);
-        Some(self.create_compaction_task(ret, ctx.base_level))
     }
 
     fn name(&self) -> &'static str {
@@ -468,7 +457,10 @@ pub mod tests {
         l0
     }
 
-    fn assert_compaction_task(compact_task: &CompactionTask, level_handlers: &[LevelHandler]) {
+    pub(crate) fn assert_compaction_task(
+        compact_task: &CompactionTask,
+        level_handlers: &[LevelHandler],
+    ) {
         for i in &compact_task.input.input_levels {
             for t in &i.table_infos {
                 assert!(level_handlers[i.level_idx as usize].is_pending_compact(&t.id));
@@ -488,7 +480,7 @@ pub mod tests {
             .compaction_mode(CompactionMode::Range as i32)
             .build();
         let selector =
-            DynamicLevelSelector::new(Arc::new(config), Arc::new(RangeOverlapStrategy::default()));
+            LevelSelectorCore::new(Arc::new(config), Arc::new(RangeOverlapStrategy::default()));
         let levels = vec![
             generate_level(1, vec![]),
             generate_level(2, generate_tables(0..5, 0..1000, 3, 10)),
@@ -634,180 +626,5 @@ pub mod tests {
         // to score.
         let compaction = selector.pick_compaction(2, &levels, &mut levels_handlers);
         assert!(compaction.is_none());
-    }
-
-    #[test]
-    fn test_manual_compaction_picker_l0() {
-        let config = Arc::new(CompactionConfigBuilder::new().max_level(4).build());
-        let selector = DynamicLevelSelector::new(config, Arc::new(RangeOverlapStrategy::default()));
-        let l0 = generate_l0_nonoverlapping_sublevels(vec![
-            generate_table(0, 1, 0, 500, 1),
-            generate_table(1, 1, 0, 500, 1),
-        ]);
-        assert_eq!(l0.sub_levels.len(), 2);
-        let levels = vec![
-            generate_level(1, vec![]),
-            generate_level(2, vec![]),
-            generate_level(3, vec![]),
-            Level {
-                level_idx: 4,
-                level_type: LevelType::Nonoverlapping as i32,
-                table_infos: vec![
-                    generate_table(2, 1, 0, 100, 1),
-                    generate_table(3, 1, 101, 200, 1),
-                    generate_table(4, 1, 222, 300, 1),
-                ],
-                total_file_size: 0,
-                sub_level_id: 0,
-            },
-        ];
-        assert_eq!(levels.len(), 4);
-        let levels = Levels {
-            levels,
-            l0: Some(l0),
-        };
-        let mut levels_handler = (0..5).into_iter().map(LevelHandler::new).collect_vec();
-
-        // pick_l0_to_sub_level
-        {
-            let option = ManualCompactionOption {
-                sst_ids: vec![0, 1],
-                key_range: KeyRange {
-                    left: vec![],
-                    right: vec![],
-                    right_exclusive: false,
-                },
-                internal_table_id: HashSet::default(),
-                level: 0,
-            };
-            let task = selector
-                .manual_pick_compaction(1, &levels, &mut levels_handler, option)
-                .unwrap();
-            assert_compaction_task(&task, &levels_handler);
-            assert_eq!(task.input.input_levels.len(), 2);
-            assert_eq!(task.input.input_levels[0].level_idx, 0);
-            assert_eq!(task.input.input_levels[1].level_idx, 0);
-            assert_eq!(task.input.target_level, 0);
-        }
-
-        for level_handler in &mut levels_handler {
-            for pending_task_id in &level_handler.pending_tasks_ids() {
-                level_handler.remove_task(*pending_task_id);
-            }
-        }
-
-        // pick_l0_to_base_level
-        {
-            let option = ManualCompactionOption {
-                sst_ids: vec![],
-                key_range: KeyRange {
-                    left: vec![],
-                    right: vec![],
-                    right_exclusive: false,
-                },
-                internal_table_id: HashSet::default(),
-                level: 0,
-            };
-            let task = selector
-                .manual_pick_compaction(2, &levels, &mut levels_handler, option)
-                .unwrap();
-            assert_compaction_task(&task, &levels_handler);
-            assert_eq!(task.input.input_levels.len(), 3);
-            assert_eq!(task.input.input_levels[0].level_idx, 0);
-            assert_eq!(task.input.input_levels[1].level_idx, 0);
-            assert_eq!(task.input.input_levels[2].level_idx, 4);
-            assert_eq!(task.input.target_level, 4);
-        }
-    }
-
-    /// tests `DynamicLevelSelector::manual_pick_compaction`
-    #[test]
-    fn test_manual_compaction_picker() {
-        let config = Arc::new(CompactionConfigBuilder::new().max_level(4).build());
-        let selector = DynamicLevelSelector::new(config, Arc::new(RangeOverlapStrategy::default()));
-        let l0 = generate_l0_nonoverlapping_sublevels(vec![]);
-        assert_eq!(l0.sub_levels.len(), 0);
-        let levels = vec![
-            generate_level(1, vec![]),
-            generate_level(2, vec![]),
-            generate_level(
-                3,
-                vec![
-                    generate_table(0, 1, 150, 151, 1),
-                    generate_table(1, 1, 250, 251, 1),
-                ],
-            ),
-            Level {
-                level_idx: 4,
-                level_type: LevelType::Nonoverlapping as i32,
-                table_infos: vec![
-                    generate_table(2, 1, 0, 100, 1),
-                    generate_table(3, 1, 101, 200, 1),
-                    generate_table(4, 1, 222, 300, 1),
-                ],
-                total_file_size: 0,
-                sub_level_id: 0,
-            },
-        ];
-        assert_eq!(levels.len(), 4);
-        let levels = Levels {
-            levels,
-            l0: Some(l0),
-        };
-        let mut levels_handler = (0..5).into_iter().map(LevelHandler::new).collect_vec();
-
-        // pick l3 -> l4
-        {
-            let option = ManualCompactionOption {
-                sst_ids: vec![0, 1],
-                key_range: KeyRange {
-                    left: vec![],
-                    right: vec![],
-                    right_exclusive: false,
-                },
-                internal_table_id: HashSet::default(),
-                level: 3,
-            };
-            let task = selector
-                .manual_pick_compaction(1, &levels, &mut levels_handler, option)
-                .unwrap();
-            assert_compaction_task(&task, &levels_handler);
-            assert_eq!(task.input.input_levels.len(), 2);
-            assert_eq!(task.input.input_levels[0].level_idx, 3);
-            assert_eq!(task.input.input_levels[0].table_infos.len(), 2);
-            assert_eq!(task.input.input_levels[1].level_idx, 4);
-            assert_eq!(task.input.input_levels[1].table_infos.len(), 2);
-            assert_eq!(task.input.target_level, 4);
-        }
-
-        for level_handler in &mut levels_handler {
-            for pending_task_id in &level_handler.pending_tasks_ids() {
-                level_handler.remove_task(*pending_task_id);
-            }
-        }
-
-        // pick l4 -> l4
-        {
-            let option = ManualCompactionOption {
-                sst_ids: vec![],
-                key_range: KeyRange {
-                    left: vec![],
-                    right: vec![],
-                    right_exclusive: false,
-                },
-                internal_table_id: HashSet::default(),
-                level: 4,
-            };
-            let task = selector
-                .manual_pick_compaction(1, &levels, &mut levels_handler, option)
-                .unwrap();
-            assert_compaction_task(&task, &levels_handler);
-            assert_eq!(task.input.input_levels.len(), 2);
-            assert_eq!(task.input.input_levels[0].level_idx, 4);
-            assert_eq!(task.input.input_levels[0].table_infos.len(), 3);
-            assert_eq!(task.input.input_levels[1].level_idx, 4);
-            assert_eq!(task.input.input_levels[1].table_infos.len(), 0);
-            assert_eq!(task.input.target_level, 4);
-        }
     }
 }
