+++ src/meta/src/hummock/compaction/manual_compaction_picker.rs
@@ -17,12 +17,18 @@ use std::sync::Arc;
 
 use itertools::Itertools;
 use risingwave_hummock_sdk::compaction_group::hummock_version_ext::HummockLevelsExt;
+use risingwave_hummock_sdk::HummockCompactionTaskId;
 use risingwave_pb::hummock::hummock_version::Levels;
-use risingwave_pb::hummock::{InputLevel, Level, LevelType, OverlappingLevel, SstableInfo};
+use risingwave_pb::hummock::{
+    CompactionConfig, InputLevel, Level, LevelType, OverlappingLevel, SstableInfo,
+};
 
 use super::overlap_strategy::OverlapInfo;
+use crate::hummock::compaction::level_selector::{LevelSelector, LevelSelectorCore};
 use crate::hummock::compaction::overlap_strategy::{OverlapStrategy, RangeOverlapInfo};
-use crate::hummock::compaction::{CompactionInput, CompactionPicker, ManualCompactionOption};
+use crate::hummock::compaction::{
+    CompactionInput, CompactionPicker, CompactionTask, ManualCompactionOption,
+};
 use crate::hummock::level_handler::LevelHandler;
 
 pub struct ManualCompactionPicker {
@@ -312,6 +318,66 @@ impl CompactionPicker for ManualCompactionPicker {
     }
 }
 
+pub struct ManualCompactionSelector {
+    inner: LevelSelectorCore,
+    option: ManualCompactionOption,
+}
+
+impl ManualCompactionSelector {
+    pub fn new(
+        config: Arc<CompactionConfig>,
+        overlap_strategy: Arc<dyn OverlapStrategy>,
+        option: ManualCompactionOption,
+    ) -> Self {
+        Self {
+            inner: LevelSelectorCore::new(config, overlap_strategy),
+            option,
+        }
+    }
+}
+
+impl LevelSelector for ManualCompactionSelector {
+    fn need_compaction(&self, levels: &Levels, _: &[LevelHandler]) -> bool {
+        let ctx = self.inner.calculate_level_base_size(levels);
+        if self.option.level > 0 && self.option.level < ctx.base_level {
+            return false;
+        }
+        true
+    }
+
+    fn pick_compaction(
+        &self,
+        task_id: HummockCompactionTaskId,
+        levels: &Levels,
+        level_handlers: &mut [LevelHandler],
+    ) -> Option<CompactionTask> {
+        let ctx = self.inner.calculate_level_base_size(levels);
+        let target_level = if self.option.level == 0 {
+            ctx.base_level
+        } else if self.option.level == self.inner.get_config().max_level as usize {
+            self.option.level
+        } else {
+            self.option.level + 1
+        };
+        if self.option.level > 0 && self.option.level < ctx.base_level {
+            return None;
+        }
+        let picker = ManualCompactionPicker::new(
+            self.inner.get_overlap_strategy(),
+            self.option.clone(),
+            target_level,
+        );
+
+        let ret = picker.pick_compaction(levels, level_handlers)?;
+        ret.add_pending_task(task_id, level_handlers);
+        Some(self.inner.create_compaction_task(ret, ctx.base_level))
+    }
+
+    fn name(&self) -> &'static str {
+        "ManualCompactionSelector"
+    }
+}
+
 #[cfg(test)]
 pub mod tests {
     use std::collections::HashSet;
@@ -319,8 +385,10 @@ pub mod tests {
     pub use risingwave_pb::hummock::{KeyRange, Level, LevelType};
 
     use super::*;
+    use crate::hummock::compaction::compaction_config::CompactionConfigBuilder;
     use crate::hummock::compaction::level_selector::tests::{
-        generate_l0_nonoverlapping_sublevels, generate_l0_overlapping_sublevels, generate_table,
+        assert_compaction_task, generate_l0_nonoverlapping_sublevels,
+        generate_l0_overlapping_sublevels, generate_level, generate_table,
     };
     use crate::hummock::compaction::overlap_strategy::RangeOverlapStrategy;
     use crate::hummock::test_utils::iterator_test_key_of_epoch;
@@ -1074,6 +1142,200 @@ pub mod tests {
                     *e
                 );
             }
+        }
+    }
+
+    #[test]
+    fn test_manual_compaction_picker_l0() {
+        let config = Arc::new(CompactionConfigBuilder::new().max_level(4).build());
+        let l0 = generate_l0_nonoverlapping_sublevels(vec![
+            generate_table(0, 1, 0, 500, 1),
+            generate_table(1, 1, 0, 500, 1),
+        ]);
+        assert_eq!(l0.sub_levels.len(), 2);
+        let levels = vec![
+            generate_level(1, vec![]),
+            generate_level(2, vec![]),
+            generate_level(3, vec![]),
+            Level {
+                level_idx: 4,
+                level_type: LevelType::Nonoverlapping as i32,
+                table_infos: vec![
+                    generate_table(2, 1, 0, 100, 1),
+                    generate_table(3, 1, 101, 200, 1),
+                    generate_table(4, 1, 222, 300, 1),
+                ],
+                total_file_size: 0,
+                sub_level_id: 0,
+            },
+        ];
+        assert_eq!(levels.len(), 4);
+        let levels = Levels {
+            levels,
+            l0: Some(l0),
+        };
+        let mut levels_handler = (0..5).into_iter().map(LevelHandler::new).collect_vec();
+
+        // pick_l0_to_sub_level
+        {
+            let option = ManualCompactionOption {
+                sst_ids: vec![0, 1],
+                key_range: KeyRange {
+                    left: vec![],
+                    right: vec![],
+                    right_exclusive: false,
+                },
+                internal_table_id: HashSet::default(),
+                level: 0,
+            };
+            let selector = ManualCompactionSelector::new(
+                config.clone(),
+                Arc::new(RangeOverlapStrategy::default()),
+                option,
+            );
+            let task = selector
+                .pick_compaction(1, &levels, &mut levels_handler)
+                .unwrap();
+            assert_compaction_task(&task, &levels_handler);
+            assert_eq!(task.input.input_levels.len(), 2);
+            assert_eq!(task.input.input_levels[0].level_idx, 0);
+            assert_eq!(task.input.input_levels[1].level_idx, 0);
+            assert_eq!(task.input.target_level, 0);
+        }
+
+        for level_handler in &mut levels_handler {
+            for pending_task_id in &level_handler.pending_tasks_ids() {
+                level_handler.remove_task(*pending_task_id);
+            }
+        }
+
+        // pick_l0_to_base_level
+        {
+            let option = ManualCompactionOption {
+                sst_ids: vec![],
+                key_range: KeyRange {
+                    left: vec![],
+                    right: vec![],
+                    right_exclusive: false,
+                },
+                internal_table_id: HashSet::default(),
+                level: 0,
+            };
+            let selector = ManualCompactionSelector::new(
+                config.clone(),
+                Arc::new(RangeOverlapStrategy::default()),
+                option,
+            );
+            let task = selector
+                .pick_compaction(2, &levels, &mut levels_handler)
+                .unwrap();
+            assert_compaction_task(&task, &levels_handler);
+            assert_eq!(task.input.input_levels.len(), 3);
+            assert_eq!(task.input.input_levels[0].level_idx, 0);
+            assert_eq!(task.input.input_levels[1].level_idx, 0);
+            assert_eq!(task.input.input_levels[2].level_idx, 4);
+            assert_eq!(task.input.target_level, 4);
+        }
+    }
+
+    /// tests `DynamicLevelSelector::manual_pick_compaction`
+    #[test]
+    fn test_manual_compaction_picker() {
+        let config = Arc::new(CompactionConfigBuilder::new().max_level(4).build());
+        let l0 = generate_l0_nonoverlapping_sublevels(vec![]);
+        assert_eq!(l0.sub_levels.len(), 0);
+        let levels = vec![
+            generate_level(1, vec![]),
+            generate_level(2, vec![]),
+            generate_level(
+                3,
+                vec![
+                    generate_table(0, 1, 150, 151, 1),
+                    generate_table(1, 1, 250, 251, 1),
+                ],
+            ),
+            Level {
+                level_idx: 4,
+                level_type: LevelType::Nonoverlapping as i32,
+                table_infos: vec![
+                    generate_table(2, 1, 0, 100, 1),
+                    generate_table(3, 1, 101, 200, 1),
+                    generate_table(4, 1, 222, 300, 1),
+                ],
+                total_file_size: 0,
+                sub_level_id: 0,
+            },
+        ];
+        assert_eq!(levels.len(), 4);
+        let levels = Levels {
+            levels,
+            l0: Some(l0),
+        };
+        let mut levels_handler = (0..5).into_iter().map(LevelHandler::new).collect_vec();
+
+        // pick l3 -> l4
+        {
+            let option = ManualCompactionOption {
+                sst_ids: vec![0, 1],
+                key_range: KeyRange {
+                    left: vec![],
+                    right: vec![],
+                    right_exclusive: false,
+                },
+                internal_table_id: HashSet::default(),
+                level: 3,
+            };
+            let selector = ManualCompactionSelector::new(
+                config.clone(),
+                Arc::new(RangeOverlapStrategy::default()),
+                option,
+            );
+            let task = selector
+                .pick_compaction(1, &levels, &mut levels_handler)
+                .unwrap();
+            assert_compaction_task(&task, &levels_handler);
+            assert_eq!(task.input.input_levels.len(), 2);
+            assert_eq!(task.input.input_levels[0].level_idx, 3);
+            assert_eq!(task.input.input_levels[0].table_infos.len(), 2);
+            assert_eq!(task.input.input_levels[1].level_idx, 4);
+            assert_eq!(task.input.input_levels[1].table_infos.len(), 2);
+            assert_eq!(task.input.target_level, 4);
+        }
+
+        for level_handler in &mut levels_handler {
+            for pending_task_id in &level_handler.pending_tasks_ids() {
+                level_handler.remove_task(*pending_task_id);
+            }
+        }
+
+        // pick l4 -> l4
+        {
+            let option = ManualCompactionOption {
+                sst_ids: vec![],
+                key_range: KeyRange {
+                    left: vec![],
+                    right: vec![],
+                    right_exclusive: false,
+                },
+                internal_table_id: HashSet::default(),
+                level: 4,
+            };
+            let selector = ManualCompactionSelector::new(
+                config,
+                Arc::new(RangeOverlapStrategy::default()),
+                option,
+            );
+
+            let task = selector
+                .pick_compaction(1, &levels, &mut levels_handler)
+                .unwrap();
+            assert_compaction_task(&task, &levels_handler);
+            assert_eq!(task.input.input_levels.len(), 2);
+            assert_eq!(task.input.input_levels[0].level_idx, 4);
+            assert_eq!(task.input.input_levels[0].table_infos.len(), 3);
+            assert_eq!(task.input.input_levels[1].level_idx, 4);
+            assert_eq!(task.input.input_levels[1].table_infos.len(), 0);
+            assert_eq!(task.input.target_level, 4);
         }
     }
 }
