+++ src/meta/src/hummock/compaction_group/mod.rs
@@ -23,7 +23,7 @@ use risingwave_pb::hummock::CompactionConfig;
 
 use crate::model::{MetadataModel, MetadataModelResult};
 
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Debug, Clone)]
 pub struct CompactionGroup {
     pub(crate) group_id: CompactionGroupId,
     pub(crate) parent_group_id: CompactionGroupId,
